requirements.txt
fastapi
uvicorn
sqlalchemy
redis
requests
python-dotenv
pydantic
scikit-learn
pandas
firebase-admin